+++ src/Utils.php
@@ -9,7 +9,7 @@ use Contentstack\Utils\Model\Metadata;
 use Contentstack\Utils\Enum\NodeType;
 use Contentstack\Utils\Enum\MarkType;
 
-class Utils
+class Utils extends BaseParser
 {
     /**
      * 
@@ -70,94 +70,21 @@ class Utils
     public static function jsonToHtml(object $content, Option $option): string {
         $resultHtml = '';
         if (isset($content->children)) {
-            $resultHtml = Utils::nodeChildrenToHtml($content->children, $option);
-        }
-        return $resultHtml;
-    }
-
-    private static function nodeChildrenToHtml(array $nodes, Option $option): string {
-        return \implode('', \array_map(function (object $node) use ($option): string {
-            return Utils::nodeToHtml($node, $option);
-        }, $nodes));
-    }
-
-    private static function nodeToHtml(object $node, Option $option): string {
-        $resultHtml = '';
-        if (isset($node->type)) {
-            switch ($node->type) {
-                case NodeType::get(NodeType::REFERENCE)->getValue():
-                    $resultHtml = Utils::referenceToHtml($node, $option);
-                    break;
-                default:
-                    $innerHtml = "";
-                    if (isset($node->children)) 
-                    {
-                        $innerHtml = Utils::nodeChildrenToHtml($node->children, $option);
+            $resultHtml = Utils::nodeChildrenToHtml($content->children, $option, function (Metadata $metadata) use ($option): string {
+                $resultHtml = '';
+                if ($option->entry) {
+                    $object = Utils::findObject($metadata, $option->entry);   
+                    if (count($object) > 0) {
+                        $resultHtml = $option->renderOptions($object, $metadata);
                     }
-                    $resultHtml = $option->renderNode(
-                        $node->type, 
-                        $node, 
-                        $innerHtml
-                    );
-                    break;
-            }
-        } else {
-            $resultHtml = Utils::textToHtml($node, $option);
+                }
+                return $resultHtml;
+            });
         }
         return $resultHtml;
     }
 
-    private static function textToHtml(object $node, Option $option) 
-    {
-        $text = $node->text;
-        if (isset($node->superscript) && $node->superscript) {
-            $text = $option->renderMark(MarkType::get(MarkType::SUPERSCRIPT), $text);
-        }
-        if (isset($node->subscript) && $node->subscript) {
-            $text = $option->renderMark(MarkType::get(MarkType::SUBSCRIPT), $text);
-        }
-        if (isset($node->inlineCode) && $node->inlineCode) {
-            $text = $option->renderMark(MarkType::get(MarkType::INLINE_CODE), $text);
-        }
-        if (isset($node->strikethrough) && $node->strikethrough) {
-            $text = $option->renderMark(MarkType::get(MarkType::STRIKE_THROUGH), $text);
-        }
-        if (isset($node->underline) && $node->underline) {
-            $text = $option->renderMark(MarkType::get(MarkType::UNDERLINE), $text);
-        }
-        if (isset($node->italic) && $node->italic) {
-            $text = $option->renderMark(MarkType::get(MarkType::ITALIC), $text);
-        }
-        if (isset($node->bold) && $node->bold) {
-            $text = $option->renderMark(MarkType::get(MarkType::BOLD), $text);
-        }
-        return $text;
-    }
-
-    private static function referenceToHtml(object $node, Option $option) 
-    {
-        $resultHtml = '';
-        if ($option->entry) {
-            $metadata = new Metadata($node);
-            $object = Utils::findObject($metadata, $option->entry);
-            if (count($object) > 0) {
-                $resultHtml = $option->renderOptions($object, $metadata);
-            }
-        }
-        return $resultHtml;
-    }
-
-    private static function findEmbeddedObject(\DOMDocument $doc): array {
-        $xpath = new \DOMXPath($doc);
-        $elements = $xpath->query('//*[contains(@class, "embedded-asset") or contains(@class, "embedded-entry")]');
-        $metadataArray = array();
-        foreach ($elements as $node) {
-            $metadataArray[] = new Metadata($node);
-        }
-        return $metadataArray;
-    }
-
-    private static function findObject(Metadata $metadata, array $entry): array
+    protected static function findObject(Metadata $metadata, array $entry): array
     {
         if (array_key_exists('_embedded_items', $entry)) 
         {
@@ -174,14 +101,4 @@ class Utils
         }
         return [];
     }
-
-    static function innerHTML(\DOMElement $element) 
-    { 
-        $doc = $element->ownerDocument;
-        $html = '';
-        foreach ($element->childNodes as $node) {
-            $html .= $doc->saveHTML($node);
-        }
-        return $html;
-    } 
 }