+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.0](https://github.com/contentstack/contentstack-utils-php/tree/v1.1.0) (2021-07-16)
+  - JSON RTE to html content functionality for GQL API added
+## [1.0.1](https://github.com/contentstack/contentstack-utils-php/tree/v1.0.1) (2021-07-16)
+  - JSON RTE to html content functionality added
 
 ## [1.0.0](https://github.com/contentstack/contentstack-utils-php/tree/v1.0.0) (2021-04-05)
   - Initial release for Contentstack Utils SDK